Change MixLine to have plate, wells, names, and do formatting
Formatting of multi-component mixlines was previously done by each
action providing strings.  This moves formatting to a location
property on the MixLine, and a list of names on the MixLine, which it
then formats.

This has the slight potential for the bold formatting to become
confused, but addressing it would likely be harder than justified for
the slight formatting irregularities in a currently unknown edge case.