Change MixLine to have plate, wells, names, and do formatting (#10)
* Change MixLine to have plate, wells, names, and do formatting

Formatting of multi-component mixlines was previously done by each
action providing strings.  This moves formatting to a location
property on the MixLine, and a list of names on the MixLine, which it
then formats.

This has the slight potential for the bold formatting to become
confused, but addressing it would likely be harder than justified for
the slight formatting irregularities in a currently unknown edge case.

Previously, the "Total" line for mixes were handled by having the mix
make those lines italic.  This change puts a `fake` parameter on
MixLine, which will make it italic, and exclude it from any
calculations done on the mix lines.